requirements.txt
pypdf==6.7.5
pyyaml==6.0.3
chromadb
sentence_transformers
openai